backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
numpy>=1.24.0
pydantic>=2.4.0
requests>=2.28.0
pymysql>=1.1.0
python-dotenv>=1.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-mock>=3.10.0
psutil>=5.9.5 